Extract layers computation
