+++ packages/alfa-rules/src/common/dom/get-colors.ts
@@ -4,3 +4,4 @@
 export * from "./get-colors/color";
 export * from "./get-colors/color-error";
 export * from "./get-colors/get-colors";
+export * from "./get-colors/layers";
